Browser Buffer data structure added
The browser buffer is an array of only the entries being displayed
currently in the browser
- prevents having to search through the entire entry list even when
  scrolling within the current browser context

Also made list a singly linked stack that points at the head because
push and pop are most likely to occur at the bottom of the list (i.e.
the most recent entries) and reverse search was not useful

Additionally the list now points to intermediate nodes which point to
the entry themselves. This was done so copies of lists could be made
(e.g. filtered lists) without copying all the entries themselves