Browser Scrolling
Browser now handles 'j' and 'k' to scroll printed entries up and down
respectively. Press 'q' to quit.

The browser entries are now represented in a struct of type `browser_t`
that contains pointers to the nodes in the main entry list for where the
browser should start and end and which entry is curently selected.
Assumes the most recent entry is the starting point (i.e. the tail of
the linked list).