Introudced robust amount printing with amnt_to_str
Large amounts that do not fit in the length specified and are greater
than a thousand will now swap format to K suffix. Did not create M
suffix because I will never be a millionaire and I don't care to provide
support for folks who are.

With this change reduced SUM_AMOUNT_STR_LEN to 8 to add more info to
smaller summary screen

Changed default output for 0s to be "--- " to reduce clutter. No more
overflow either. If string is larger or less than what can fit in
passed max_len, "MAX " or "MIN " will be returned respectively.